+++ src/TitleDetails/PiecesList/PiecesList.js
@@ -1,4 +1,4 @@
-import React, { useCallback, useMemo } from 'react';
+import React, { useMemo } from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 
@@ -61,7 +61,6 @@ const formatter = {
 };
 
 const PiecesList = ({ pieces, id, visibleColumns, selectPiece }) => {
-  const onRowClick = useCallback((e, piece) => selectPiece && selectPiece(piece), [selectPiece]);
   const hasRowClick = Boolean(selectPiece);
   const rowProps = useMemo(() => ({ alignLastColToEnd: hasRowClick }), [hasRowClick]);
 
@@ -75,7 +74,7 @@ const PiecesList = ({ pieces, id, visibleColumns, selectPiece }) => {
       rowFormatter={acqRowFormatter}
       rowProps={rowProps}
       visibleColumns={visibleColumns}
-      onRowClick={selectPiece ? onRowClick : undefined}
+      onRowClick={selectPiece}
     />
   );
 };